API Reference

SparseVariables.SparseArray — Type

SparseArray{T,N, K <: NTuple{N,Any} }

Implementation of an AbstractSparseArray where data is stored in a dictionary.

SparseVariables.IndexedVarArray — Type

IndexedVarArray{V,N,T}

Structure for holding an optimization variable with a sparse structure with extra indexing

SparseVariables.insertvar! — Function

insertvar!(var::IndexedVarArray{V,N,T}, index...)

Insert a new variable with the given index only after checking if keys are valid and not already defined.

SparseVariables.unsafe_insertvar! — Function

unsafe_insertvar!(var::indexedVarArray{V,N,T}, index...)

Insert a new variable with the given index withouth checking if the index is valid or already assigned.